Add new result type for skin "Unchanged"
This checks to see if the provided block is the same as the existing
one, which prevents it from rebuilding the structure when it doesn't
need to.

+++ src/main/kotlin/dev/mizarc/waystonewarps/application/actions/management/UpdateWarpSkin.kt
@@ -14,6 +14,7 @@ class UpdateWarpSkin(private val warpRepository: WarpRepository,
     fun execute(warpId: UUID, blockName: String): UpdateWarpSkinResult {
         val warp = warpRepository.getById(warpId) ?: return UpdateWarpSkinResult.WARP_NOT_FOUND
         if (blockName !in allowedBlockNames) return UpdateWarpSkinResult.BLOCK_NOT_VALID
+        if (blockName == warp.block) return UpdateWarpSkinResult.UNCHANGED
         warp.block = blockName
         warpRepository.update(warp)
         structureBuilderService.destroyStructure(warp)
